fix books api
