+++ db/queries/tags/index.ts
@@ -4,7 +4,7 @@ export const getTagIdByName = (name: string) => {
   return knex.select('id').from('tags').where('name', name);
 };
 
-export const add = (name: string) => {
+export const addTag = (name: string) => {
   return knex('tags').returning('id').insert({ name });
 };
 
@@ -13,16 +13,10 @@ export const getAllTags = () => {
     `
       SELECT t.id, t.name AS label, COUNT(tr.id) as count
         FROM tags as t
-        JOIN user_tag_book as tr
+      JOIN user_tag_book as tr
         ON tr.tag_id = t.id
-        GROUP BY t.id
-        ORDER BY t.name
-      `
+      GROUP BY t.id
+      ORDER BY t.name
+    `
   );
-};
-
-export default {
-  getTagIdByName,
-  add,
-  getAllTags,
 };
