+++ pages/api/books/new.ts
@@ -1,4 +1,4 @@
-const queries = require('../../../db/queries/books');
+import { addBook, confirmBook } from '../../../db/queries/books';
 import auth0 from '../../../lib/auth0';
 import userProfileFetcher from '../../../src/helpers/userProfileFetcher';
 import { NextApiRequest, NextApiResponse } from 'next';
@@ -47,7 +47,7 @@ export const newBook = async (req: NextApiRequest, res: NextApiResponse) => {
       };
 
       try {
-        const response = await queries.books.confirm(bookObj);
+        const response = await confirmBook(bookObj);
         return res.status(200).json(response);
       } catch (error) {
         return res.status(500).json(error);
@@ -64,7 +64,7 @@ export const newBook = async (req: NextApiRequest, res: NextApiResponse) => {
       bookObj = req.body;
 
       try {
-        const response = await queries.books.add(bookObj);
+        const response = await addBook(bookObj);
         return res.status(200).json(response);
       } catch (error) {
         return res.status(500).json(error);
